Refactor code for broader review!
